[FIX] uuid and deletion
