+++ xe_customs/models/__init__.py
@@ -2,7 +2,7 @@
 # © 2024 Morwi Encoders Consulting SA DE CV
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
 
-from . import account_invoice
+from . import account_move
 from . import product_clientinfo
 from . import product_product
 from . import product_template
