+++ xe_customs/models/sale_order.py
@@ -205,8 +205,8 @@ class SaleDownPayment(models.Model):
         for payment in self:
             payment.order_line_id.invoice_lines.write({
                 'sale_line_ids': False,
-                'price_unit': 0,
-                'name': _('Deleted Down Payment'),
+                # 'price_unit': 0,
+                'name': _('Unlinked Down Payment'),
             })
             payment.order_line_id.write({
                 'invoice_lines': False,
