+++ xe_customs/wizard/__init__.py
@@ -2,4 +2,5 @@
 # © 2024 Morwi Encoders Consulting SA DE CV
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
 
+from . import account_move_send
 from . import sale_make_invoice_advance
